backend/requirements.txt
fastapi
uvicorn
pathway[all]
requests
yfinance
sentence-transformers
anthropic
python-dotenv
websockets